pyproject.toml
[tool.poetry]
name = "sys-vars"
version = "1.0.0"
description = "Access system variables in your code as native Python data types."
authors = ["Caleb Ely <le717@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
python-dotenv = "^0.10.5"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
